Move some code around. IntelliJ code formatter run
